Fix pre setup scripts to ensure correct return codes.

+++ vim/_pre_remove.sh
@@ -9,4 +9,6 @@
 
 # Remove all dirs found in the bundle dir
 BUNDLE=${INSTALLTARGET}/.vim/bundle
-find $BUNDLE -maxdepth 1 ! -path $BUNDLE -type d -exec rm -rf {} \;
+if [[ -d $BUNDLE ]]; then
+    find $BUNDLE -maxdepth 1 ! -path $BUNDLE -type d -exec rm -rf {} \;
+fi
